#50 refactor(archive): 보관함 페이지 편지 상세 뷰를 동적 라우트로 분리
- 편지 상세 페이지를 archive/[id]/page.tsx로 분리하여 동적 라우팅 적용
- LetterListItem 컴포넌트를 components/archive/로 분리하여 재사용성 향상
- archive/page.tsx를 편지 리스트 전용 페이지로 단순화
- 불필요한 import 및 상태 관리 로직 제거
